+++ geo/src/algorithm/triangulate_delaunay.rs
@@ -61,36 +61,43 @@ impl<T: GeoFloat + SpadeNum> SpadeTriangulationFloat for T {}
 
 pub type Triangles<T> = Vec<Triangle<T>>;
 
-// seal the trait that needs to be implemented for TriangulateDelaunay to be implemented. This is done
-// so that we don't leak these weird methods on the public interface.
+// Sealed traits for triangulation requirements. Split into two:
+// - UnconstrainedRequirementTrait: only needs coords (for point collections)
+// - ConstrainedRequirementTrait: needs coords + lines + contains (for polygons etc.)
 mod private {
     use super::*;
 
     pub(crate) type CoordsIter<'a, T> = Box<dyn Iterator<Item = Coord<T>> + 'a>;
 
-    pub trait TriangulationRequirementTrait<'a, T>
+    /// Requirement trait for unconstrained triangulation: only needs coordinates
+    pub trait UnconstrainedRequirementTrait<'a, T>
     where
         T: SpadeTriangulationFloat,
     {
-        /// collect all the lines that are relevant for triangulations from the geometric object that
-        /// should be triangulated.
-        ///
-        /// intersecting lines are allowed
-        fn lines(&'a self) -> Vec<Line<T>>;
-        /// collect all the coords that are relevant for triangulations from the geometric object that
-        /// should be triangulated
+        /// Collect all coords from the geometry
         fn coords(&'a self) -> CoordsIter<'a, T>;
-        /// define a predicate that decides if a point is inside of the object (used for constrained triangulation)
+    }
+
+    /// Requirement trait for constrained triangulation: needs coords, lines, and containment
+    pub trait ConstrainedRequirementTrait<'a, T>: UnconstrainedRequirementTrait<'a, T>
+    where
+        T: SpadeTriangulationFloat,
+    {
+        /// Collect all lines from the geometry (intersecting lines are allowed)
+        fn lines(&'a self) -> Vec<Line<T>>;
+
+        /// Check if a point is inside the geometry (used for constrained triangulation)
         fn contains_point(&'a self, p: Point<T>) -> bool;
 
-        // processing of the lines that prepare the lines for triangulation.
+        // Processing of the lines that prepare the lines for triangulation.
         //
-        // `spade` has the general limitation that constraint lines cannot intersect or else it
-        // will panic. This is why we need to manually split up the lines into smaller parts at the
-        // intersection point
+        // Spade has the limitation that constraint lines cannot intersect or else it
+        // will panic. This is why we need to manually split up the lines into smaller
+        // parts at the intersection point.
         //
-        // there's also a preprocessing step which tries to minimize the risk of failure of the algo
-        // through edge cases (thin/flat triangles are prevented as much as possible & lines are deduped, ...)
+        // There's also a preprocessing step which tries to minimize the risk of failure
+        // through edge cases (thin/flat triangles are prevented as much as possible,
+        // lines are deduped, etc.)
         fn cleanup_lines(lines: Vec<Line<T>>, snap_radius: T) -> TriangulationResult<Vec<Line<T>>> {
             let (known_coords, lines) = preprocess_lines(lines, snap_radius);
             prepare_intersection_contraint(lines, known_coords, snap_radius)
@@ -98,45 +105,38 @@ mod private {
     }
 }
 
-/// Triangulate polygons using a [Delaunay
-/// Triangulation](https://en.wikipedia.org/wiki/Delaunay_triangulation)
+/// Unconstrained Delaunay triangulation for geometries that provide coordinates.
 ///
-/// This trait contains both constrained and unconstrained triangulation methods. To read more
-/// about the differences of these methods also consult [this
-/// page](https://en.wikipedia.org/wiki/Constrained_Delaunay_triangulation)
-pub trait TriangulateDelaunay<'a, T>: private::TriangulationRequirementTrait<'a, T>
+/// This trait provides triangulation methods based only on point coordinates,
+/// making it suitable for point collections like `MultiPoint`, `Vec<Point>`, etc.
+///
+/// For constrained triangulation methods that respect geometries' existing boundaries,
+/// see [`TriangulateDelaunay`].
+pub trait TriangulateDelaunayUnconstrained<'a, T>:
+    private::UnconstrainedRequirementTrait<'a, T>
 where
     T: SpadeTriangulationFloat,
 {
-    /// returns a triangulation that's solely based on the points of the geometric object
+    /// Returns a triangulation based solely on the points of the geometric object.
     ///
-    /// The triangulation is guaranteed to be Delaunay
+    /// The triangulation is guaranteed to be Delaunay.
     ///
     /// Note that the lines of the triangulation don't necessarily follow the lines of the input
-    /// geometry. If you wish to achieve that take a look at the `constrained_triangulation` and the
-    /// `constrained_outer_triangulation` functions.
+    /// geometry. If you wish to achieve that, see [`TriangulateDelaunay::constrained_triangulation`]
+    /// and [`TriangulateDelaunay::constrained_outer_triangulation`].
     ///
     /// ```rust
-    /// use geo::TriangulateDelaunay;
-    /// use geo::{Polygon, LineString, Coord};
-    /// let u_shape = Polygon::new(
-    ///     LineString::new(vec![
-    ///         Coord { x: 0.0, y: 0.0 },
-    ///         Coord { x: 1.0, y: 0.0 },
-    ///         Coord { x: 1.0, y: 1.0 },
-    ///         Coord { x: 2.0, y: 1.0 },
-    ///         Coord { x: 2.0, y: 0.0 },
-    ///         Coord { x: 3.0, y: 0.0 },
-    ///         Coord { x: 3.0, y: 3.0 },
-    ///         Coord { x: 0.0, y: 3.0 },
-    ///     ]),
-    ///     vec![],
-    /// );
-    /// let unconstrained_triangulation = u_shape.unconstrained_triangulation().unwrap();
-    /// let num_triangles = unconstrained_triangulation.len();
-    /// assert_eq!(num_triangles, 8);
-    /// ```
+    /// use geo::TriangulateDelaunayUnconstrained;
+    /// use geo::{MultiPoint, Point};
     ///
+    /// let points = MultiPoint::new(vec![
+    ///     Point::new(0.0, 0.0),
+    ///     Point::new(1.0, 0.0),
+    ///     Point::new(0.5, 1.0),
+    /// ]);
+    /// let triangulation = points.unconstrained_triangulation().unwrap();
+    /// assert_eq!(triangulation.len(), 1);
+    /// ```
     fn unconstrained_triangulation(&'a self) -> TriangulationResult<Triangles<T>> {
         let points = self.coords();
         points
@@ -148,11 +148,22 @@ where
             })
             .map(triangulation_to_triangles)
     }
+}
 
-    /// returns triangulation that's based on the points of the geometric object and also
-    /// incorporates the lines of the input geometry
+/// Triangulate polygons using a [Delaunay Triangulation](https://en.wikipedia.org/wiki/Delaunay_triangulation)
+///
+/// This trait extends [`TriangulateDelaunayUnconstrained`] with constrained triangulation methods
+/// that respect geometry boundaries. To read more about the differences between constrained and
+/// unconstrained methods, see the [Wikipedia article on constrained Delaunay triangulation](https://en.wikipedia.org/wiki/Constrained_Delaunay_triangulation).
+pub trait TriangulateDelaunay<'a, T>:
+    TriangulateDelaunayUnconstrained<'a, T> + private::ConstrainedRequirementTrait<'a, T>
+where
+    T: SpadeTriangulationFloat,
+{
+    /// Returns triangulation that's based on the points of the geometric object and also
+    /// incorporates the lines of the input geometry.
     ///
-    /// The triangulation is not guaranteed to be Delaunay because of the constraint lines
+    /// The triangulation is not guaranteed to be Delaunay because of the constraint lines.
     ///
     /// This outer triangulation also contains triangles that are not included in the input
     /// geometry if it wasn't convex. Here's an example:
@@ -223,10 +234,10 @@ where
             .map(triangulation_to_triangles)
     }
 
-    /// returns triangulation that's based on the points of the geometric object and also
-    /// incorporates the lines of the input geometry
+    /// Returns triangulation that's based on the points of the geometric object and also
+    /// incorporates the lines of the input geometry.
     ///
-    /// The triangulation is not guaranteed to be Delaunay because of the constraint lines
+    /// The triangulation is not guaranteed to be Delaunay because of the constraint lines.
     ///
     /// This triangulation only contains triangles that are included in the input geometry.
     /// Here's an example:
@@ -304,24 +315,40 @@ where
 
 // ========== Triangulation trait impls ============
 
-// everything that satisfies the requirement methods automatically implements the triangulation
-impl<'a, T, G> TriangulateDelaunay<'a, T> for G
+// Blanket impl: anything with UnconstrainedRequirementTrait gets TriangulateDelaunayUnconstrained
+impl<'a, T, G> TriangulateDelaunayUnconstrained<'a, T> for G
 where
     T: SpadeTriangulationFloat,
-    G: private::TriangulationRequirementTrait<'a, T>,
+    G: private::UnconstrainedRequirementTrait<'a, T>,
 {
 }
 
-impl<'a, 'l, T, G> private::TriangulationRequirementTrait<'a, T> for G
+// Blanket impl: anything with ConstrainedRequirementTrait gets TriangulateDelaunay
+impl<'a, T, G> TriangulateDelaunay<'a, T> for G
+where
+    T: SpadeTriangulationFloat,
+    G: private::ConstrainedRequirementTrait<'a, T>,
+{
+}
+
+// Impl UnconstrainedRequirementTrait for types with CoordsIter
+impl<'a, T, G> private::UnconstrainedRequirementTrait<'a, T> for G
+where
+    T: SpadeTriangulationFloat,
+    G: CoordsIter<Scalar = T>,
+{
+    fn coords(&'a self) -> private::CoordsIter<'a, T> {
+        Box::new(self.coords_iter())
+    }
+}
+
+// Impl ConstrainedRequirementTrait for types with LinesIter + CoordsIter + Contains
+impl<'a, 'l, T, G> private::ConstrainedRequirementTrait<'a, T> for G
 where
     'a: 'l,
     T: SpadeTriangulationFloat,
     G: LinesIter<'l, Scalar = T> + CoordsIter<Scalar = T> + Contains<Point<T>>,
 {
-    fn coords(&'a self) -> private::CoordsIter<'a, T> {
-        Box::new(self.coords_iter())
-    }
-
     fn lines(&'a self) -> Vec<Line<T>> {
         self.lines_iter().collect()
     }
@@ -331,18 +358,21 @@ where
     }
 }
 
-// it would be cool to impl the trait for GS: AsRef<[G]> but I wasn't able to get this to compile
-// (yet)
-
-impl<'a, T, G> private::TriangulationRequirementTrait<'a, T> for Vec<G>
+impl<'a, T, G> private::UnconstrainedRequirementTrait<'a, T> for Vec<G>
 where
     T: SpadeTriangulationFloat + 'a,
-    G: TriangulateDelaunay<'a, T>,
+    G: TriangulateDelaunayUnconstrained<'a, T>,
 {
     fn coords(&'a self) -> private::CoordsIter<'a, T> {
         Box::new(self.iter().flat_map(|g| g.coords()))
     }
+}
 
+impl<'a, T, G> private::ConstrainedRequirementTrait<'a, T> for Vec<G>
+where
+    T: SpadeTriangulationFloat + 'a,
+    G: TriangulateDelaunay<'a, T>,
+{
     fn lines(&'a self) -> Vec<Line<T>> {
         self.iter().flat_map(|g| g.lines()).collect::<Vec<_>>()
     }
@@ -352,15 +382,25 @@ where
     }
 }
 
-impl<'a, T, G> private::TriangulationRequirementTrait<'a, T> for &[G]
+// Note: We add LinesIter bound to disambiguate from &[Coord] which implements CoordsIter.
+// Coord doesn't implement LinesIter, so this prevents the coherence conflict.
+impl<'a, 'l, T, G> private::UnconstrainedRequirementTrait<'a, T> for &[G]
 where
+    'a: 'l,
     T: SpadeTriangulationFloat + 'a,
-    G: TriangulateDelaunay<'a, T>,
+    G: TriangulateDelaunayUnconstrained<'a, T> + LinesIter<'l, Scalar = T>,
 {
     fn coords(&'a self) -> private::CoordsIter<'a, T> {
         Box::new(self.iter().flat_map(|g| g.coords()))
     }
+}
 
+impl<'a, 'l, T, G> private::ConstrainedRequirementTrait<'a, T> for &[G]
+where
+    'a: 'l,
+    T: SpadeTriangulationFloat + 'a,
+    G: TriangulateDelaunay<'a, T> + LinesIter<'l, Scalar = T>,
+{
     fn lines(&'a self) -> Vec<Line<T>> {
         self.iter().flat_map(|g| g.lines()).collect::<Vec<_>>()
     }
@@ -369,6 +409,11 @@ where
         self.iter().any(|g| g.contains_point(p))
     }
 }
+
+// Note: Point collections like MultiPoint, Point, and &[Coord] get
+// TriangulateDelaunayUnconstrained automatically via the blanket impl for CoordsIter.
+// They only support unconstrained triangulation because points don't have edges
+// or an interior: constrained triangulation requires boundaries to constrain against.
 
 // ========== Triangulation trait impl helpers ============
 
@@ -783,5 +828,60 @@ mod spade_triangulation {
                 });
             assert_num_triangles(&constrained_triangulation, 6);
         }
+    }
+
+    #[test]
+    fn multipoint_triangulates_unconstrained() {
+        use super::TriangulateDelaunayUnconstrained;
+
+        let points: MultiPoint<f64> = wkt!(MULTIPOINT(0. 0., 1. 0., 0.5 1., 0.5 0.5));
+
+        let triangulation = points.unconstrained_triangulation();
+        // 4 points with one inside the triangle formed by the others produce 3 triangles
+        assert_num_triangles(&triangulation, 3);
+    }
+
+    #[test]
+    fn point_triangulates_unconstrained() {
+        use super::TriangulateDelaunayUnconstrained;
+
+        // Single point produces no triangles (need at least 3 points)
+        let point: Point<f64> = wkt!(POINT(0. 0.));
+        let triangulation = point.unconstrained_triangulation();
+        assert_num_triangles(&triangulation, 0);
+    }
+
+    #[test]
+    fn coord_slice_triangulates_unconstrained() {
+        use super::TriangulateDelaunayUnconstrained;
+
+        let coords: &[Coord<f64>] = &[
+            Coord { x: 0.0, y: 0.0 },
+            Coord { x: 1.0, y: 0.0 },
+            Coord { x: 0.5, y: 1.0 },
+        ];
+        let triangulation = coords.unconstrained_triangulation();
+        assert_num_triangles(&triangulation, 1);
+    }
+
+    #[test]
+    fn polygon_slice_triangulates() {
+        use super::{TriangulateDelaunay, TriangulateDelaunayUnconstrained};
+
+        let polygons: Vec<Polygon<f64>> = vec![
+            wkt!(POLYGON((0. 0., 1. 0., 0.5 1., 0. 0.))),
+            wkt!(POLYGON((2. 0., 3. 0., 2.5 1., 2. 0.))),
+        ];
+        let slice: &[Polygon<f64>] = &polygons;
+
+        // Slice of polygons should support unconstrained triangulation
+        // 6 points from 2 triangular polygons -> 4 triangles in convex hull
+        let triangulation = slice.unconstrained_triangulation();
+        assert_num_triangles(&triangulation, 4);
+
+        // Slice of polygons should support constrained triangulation
+        // Respects polygon boundaries -> 2 triangles (one per polygon)
+        let constrained = slice.constrained_triangulation(Default::default());
+        assert_num_triangles(&constrained, 2);
     }
 }